+++ deployment/ccip/changeset/v1_6/cs_active_candidate_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package v1_6_test
 
 import (
