+++ deployment/ccip/changeset/v1_6/cs_add_new_chain_e2e_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package v1_6_test
 
 import (
