+++ deployment/ccip/changeset/v1_6/cs_update_bidirectional_lanes_unit_test.go
@@ -7,22 +7,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
-	chain_selectors "github.com/smartcontractkit/chain-selectors"
 	"github.com/smartcontractkit/mcms/types"
 	"github.com/stretchr/testify/require"
 
-	cldf_chain "github.com/smartcontractkit/chainlink-deployments-framework/chain"
-
 	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/v1_6_3/fee_quoter"
 
-	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
-
-	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/testhelpers"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/v1_6"
-	"github.com/smartcontractkit/chainlink/deployment/ccip/shared/stateview"
-	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
-	commoncs "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/common/proposalutils"
 )
 
@@ -63,95 +53,6 @@ func getAllPossibleLanes(chains []v1_6.ChainDefinition, disable bool) []v1_6.Bid
 	}
 
 	return lanes
-}
-
-func getRemoteChains(chains []v1_6.ChainDefinition, currentIndex int) []v1_6.ChainDefinition {
-	remoteChains := make([]v1_6.ChainDefinition, 0, len(chains)-1)
-	for i, chain := range chains {
-		if i == currentIndex {
-			continue
-		}
-		remoteChains = append(remoteChains, chain)
-	}
-	return remoteChains
-}
-
-func checkBidirectionalLaneConnectivity(
-	t *testing.T,
-	e cldf.Environment,
-	state stateview.CCIPOnChainState,
-	chainOne v1_6.ChainDefinition,
-	chainTwo v1_6.ChainDefinition,
-	testRouter bool,
-	disable bool,
-) {
-	lanes := []laneDefinition{
-		{
-			Source: chainOne,
-			Dest:   chainTwo,
-		},
-		{
-			Source: chainTwo,
-			Dest:   chainOne,
-		},
-	}
-	for _, lane := range lanes {
-		onRamp := state.Chains[lane.Source.Selector].OnRamp
-		offRamp := state.Chains[lane.Dest.Selector].OffRamp
-		feeQuoterOnSrc := state.Chains[lane.Source.Selector].FeeQuoter
-		routerOnSrc := state.Chains[lane.Source.Selector].Router
-		routerOnDest := state.Chains[lane.Dest.Selector].Router
-		if testRouter {
-			routerOnSrc = state.Chains[lane.Source.Selector].TestRouter
-			routerOnDest = state.Chains[lane.Dest.Selector].TestRouter
-		}
-
-		destChainConfig, err := onRamp.GetDestChainConfig(nil, lane.Dest.Selector)
-		require.NoError(t, err, "must get dest chain config from onRamp")
-		routerAddr := routerOnSrc.Address().Hex()
-		if disable {
-			routerAddr = common.HexToAddress("0x0").Hex()
-		}
-		require.Equal(t, routerAddr, destChainConfig.Router.Hex(), "router must equal expected")
-		require.Equal(t, lane.Dest.AllowListEnabled, destChainConfig.AllowlistEnabled, "allowListEnabled must equal expected")
-
-		srcChainConfig, err := offRamp.GetSourceChainConfig(nil, lane.Source.Selector)
-		require.NoError(t, err, "must get src chain config from offRamp")
-		require.Equal(t, !disable, srcChainConfig.IsEnabled, "isEnabled must be expected")
-		require.Equal(t, lane.Source.RMNVerificationDisabled, srcChainConfig.IsRMNVerificationDisabled, "rmnVerificationDisabled must equal expected")
-		require.Equal(t, common.LeftPadBytes(state.Chains[lane.Source.Selector].OnRamp.Address().Bytes(), 32), srcChainConfig.OnRamp, "remote onRamp must be set on offRamp")
-		require.Equal(t, routerOnDest.Address().Hex(), srcChainConfig.Router.Hex(), "router must equal expected")
-
-		isOffRamp, err := routerOnSrc.IsOffRamp(nil, lane.Dest.Selector, state.Chains[lane.Source.Selector].OffRamp.Address())
-		require.NoError(t, err, "must check if router has offRamp")
-		require.Equal(t, !disable, isOffRamp, "isOffRamp result must equal expected")
-		onRampOnRouter, err := routerOnSrc.GetOnRamp(nil, lane.Dest.Selector)
-		require.NoError(t, err, "must get onRamp from router")
-		onRampAddr := state.Chains[lane.Source.Selector].OnRamp.Address().Hex()
-		if disable {
-			onRampAddr = common.HexToAddress("0x0").Hex()
-		}
-		require.Equal(t, onRampAddr, onRampOnRouter.Hex(), "onRamp must equal expected")
-
-		isOffRamp, err = routerOnDest.IsOffRamp(nil, lane.Source.Selector, state.Chains[lane.Dest.Selector].OffRamp.Address())
-		require.NoError(t, err, "must check if router has offRamp")
-		require.Equal(t, !disable, isOffRamp, "isOffRamp result must equal expected")
-		onRampOnRouter, err = routerOnDest.GetOnRamp(nil, lane.Source.Selector)
-		require.NoError(t, err, "must get onRamp from router")
-		onRampAddr = state.Chains[lane.Dest.Selector].OnRamp.Address().Hex()
-		if disable {
-			onRampAddr = common.HexToAddress("0x0").Hex()
-		}
-		require.Equal(t, onRampAddr, onRampOnRouter.Hex(), "onRamp must equal expected")
-
-		feeQuoterDestConfig, err := feeQuoterOnSrc.GetDestChainConfig(nil, lane.Dest.Selector)
-		require.NoError(t, err, "must get dest chain config from feeQuoter")
-		require.Equal(t, lane.Dest.FeeQuoterDestChainConfig, feeQuoterDestConfig, "feeQuoter dest chain config must equal expected")
-
-		price, err := feeQuoterOnSrc.GetDestinationChainGasPrice(nil, lane.Dest.Selector)
-		require.NoError(t, err, "must get price from feeQuoter")
-		require.Equal(t, lane.Dest.GasPrice, price.Value, "price must equal expected")
-	}
 }
 
 func TestBuildConfigs(t *testing.T) {
@@ -266,140 +167,4 @@ func TestBuildConfigs(t *testing.T) {
 		},
 		MCMS: cfg.MCMSConfig,
 	}, configs.UpdateRouterRampsConfig)
-}
-
-func TestUpdateBidirectionalLanesChangeset(t *testing.T) {
-	t.Parallel()
-
-	type test struct {
-		Msg        string
-		TestRouter bool
-		MCMS       *proposalutils.TimelockConfig
-		Disable    bool
-	}
-
-	mcmsConfig := &proposalutils.TimelockConfig{
-		MinDelay:   0 * time.Second,
-		MCMSAction: types.TimelockActionSchedule,
-	}
-
-	tests := []test{
-		{
-			Msg:        "Use production router (with MCMS) & disable afterwards",
-			TestRouter: false,
-			MCMS:       mcmsConfig,
-			Disable:    true,
-		},
-		{
-			Msg:        "Use production router (with MCMS)",
-			TestRouter: false,
-			MCMS:       mcmsConfig,
-		},
-		{
-			Msg:        "Use test router (without MCMS)",
-			TestRouter: true,
-			MCMS:       nil,
-		},
-		{
-			Msg:        "Use test router (with MCMS for other contracts)",
-			TestRouter: true,
-			MCMS:       mcmsConfig,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.Msg, func(t *testing.T) {
-			deployedEnvironment, _ := testhelpers.NewMemoryEnvironment(t, func(testCfg *testhelpers.TestConfigs) {
-				testCfg.Chains = 3
-			})
-			e := deployedEnvironment.Env
-
-			state, err := stateview.LoadOnchainState(e)
-			require.NoError(t, err, "must load onchain state")
-
-			selectors := e.BlockChains.ListChainSelectors(cldf_chain.WithFamily(chain_selectors.FamilyEVM))
-
-			if test.MCMS != nil {
-				contractsToTransfer := make(map[uint64][]common.Address, len(selectors))
-				for _, selector := range selectors {
-					contractsToTransfer[selector] = []common.Address{
-						state.Chains[selector].OnRamp.Address(),
-						state.Chains[selector].OffRamp.Address(),
-						state.Chains[selector].Router.Address(),
-						state.Chains[selector].FeeQuoter.Address(),
-						state.Chains[selector].TokenAdminRegistry.Address(),
-						state.Chains[selector].RMNRemote.Address(),
-						state.Chains[selector].RMNProxy.Address(),
-						state.Chains[selector].NonceManager.Address(),
-					}
-				}
-				e, err = commonchangeset.Apply(t, e,
-					commonchangeset.Configure(
-						cldf.CreateLegacyChangeSet(commoncs.TransferToMCMSWithTimelockV2),
-						commoncs.TransferToMCMSWithTimelockConfig{
-							ContractsByChain: contractsToTransfer,
-							MCMSConfig: proposalutils.TimelockConfig{
-								MinDelay:   0 * time.Second,
-								MCMSAction: types.TimelockActionSchedule,
-							},
-						},
-					),
-				)
-				require.NoError(t, err, "must apply TransferToMCMSWithTimelock")
-			}
-
-			chains := make([]v1_6.ChainDefinition, len(selectors))
-			for i, selector := range selectors {
-				chains[i] = v1_6.ChainDefinition{
-					ConnectionConfig: v1_6.ConnectionConfig{
-						RMNVerificationDisabled: true,
-						AllowListEnabled:        false,
-					},
-					Selector:                 selector,
-					GasPrice:                 big.NewInt(1e17),
-					FeeQuoterDestChainConfig: v1_6.DefaultFeeQuoterDestChainConfig(true),
-				}
-			}
-
-			e, err = commonchangeset.Apply(t, e,
-				commonchangeset.Configure(
-					v1_6.UpdateBidirectionalLanesChangeset,
-					v1_6.UpdateBidirectionalLanesConfig{
-						TestRouter: test.TestRouter,
-						MCMSConfig: test.MCMS,
-						Lanes:      getAllPossibleLanes(chains, false),
-					},
-				),
-			)
-			require.NoError(t, err, "must apply AddBidirectionalLanesChangeset")
-
-			for i, chain := range chains {
-				remoteChains := getRemoteChains(chains, i)
-				for _, remoteChain := range remoteChains {
-					checkBidirectionalLaneConnectivity(t, e, state, chain, remoteChain, test.TestRouter, false)
-				}
-			}
-
-			if test.Disable {
-				e, err = commonchangeset.Apply(t, e,
-					commonchangeset.Configure(
-						v1_6.UpdateBidirectionalLanesChangeset,
-						v1_6.UpdateBidirectionalLanesConfig{
-							TestRouter: test.TestRouter,
-							MCMSConfig: test.MCMS,
-							Lanes:      getAllPossibleLanes(chains, true),
-						},
-					),
-				)
-				require.NoError(t, err, "must apply AddBidirectionalLanesChangeset")
-
-				for i, chain := range chains {
-					remoteChains := getRemoteChains(chains, i)
-					for _, remoteChain := range remoteChains {
-						checkBidirectionalLaneConnectivity(t, e, state, chain, remoteChain, test.TestRouter, true)
-					}
-				}
-			}
-		})
-	}
 }
