+++ deployment/ccip/changeset/v1_6_2/cs_configure_cctp_message_transmitter_proxy_unit_test.go
@@ -1,3 +1,5 @@
+//go:build !integration
+
 package v1_6_2_test
 
 import (
