+++ deployment/ccip/shared/deployergroup/deployer_group_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package deployergroup_test
 
 import (
